refactor: create auth utility to store JWT token
